+++ src/ops/transform.rs
@@ -2,7 +2,6 @@ use crate::opbasics::*;
 
 use serde_derive::{Deserialize, Serialize};
 use std::mem;
-use std::usize;
 
 #[derive(Copy, Clone, Debug, Serialize, Deserialize)]
 pub enum Rotation {
@@ -123,9 +122,9 @@ fn rotate_buffer(buf: &OpBuffer, orientation: &Orientation) -> OpBuffer {
     let mut out = if transpose {
         mem::swap(&mut width, &mut height);
         mem::swap(&mut x_step, &mut y_step);
-        OpBuffer::new(buf.height, buf.width, 3 as usize, buf.monochrome)
+        OpBuffer::new(buf.height, buf.width, 3_usize, buf.monochrome)
     } else {
-        OpBuffer::new(buf.width, buf.height, 3 as usize, buf.monochrome)
+        OpBuffer::new(buf.width, buf.height, 3_usize, buf.monochrome)
     };
 
     out.mutate_lines(
